split into util functions
